+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 1.7.0 (not released yet)
 
 * API change: Renamed `displayName` to `label`.
-* API change: The `ActionArgMeta` class renamed to `ArgMeta`.
-* API change: A provided argument specification in now placed in the `ArgProvided` class, not directly in the `ArgMeta` as before.
-* API change: Removed `LabeledValue` and `ArgValue.valueSetDisplayNames` because of a new support for an annotated value set.
+* API change: Renamed `ActionArgMeta` to `ArgMeta`, `ActionResultMeta` to `ResultMeta`, `ArgProvided` to `ArgProvidedMeta`, `ArgValue` to `ArgProvidedValue`.
+* API change: A provided argument specification in now placed in the `ArgProvidedMeta` class, not directly in the `ArgMeta` as before.
+* API change: Removed `LabeledValue` and `ArgProvidedValue.valueSetDisplayNames` because of a new support for an annotated value set.
   
 ## 1.6.0 (2019-01-11)